requirements.txt
# Core LangChain & LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0

# Vector Store
chromadb>=0.5.0

# Embeddings (optional - for OpenAI embeddings)
openai>=1.0.0

# Docker SDK
docker>=7.0.0

# GitHub Integration
PyGithub>=2.0.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Code Parsing
tree-sitter>=0.22.0
tree-sitter-python>=0.21.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Type hints
typing-extensions>=4.9.0

